Hacker Rank Intermediate
Delighted to have earned the SQL Intermediate badge on Hacker Rank! Mastering the intricacies of SQL is not just a goal; it's a journey, and this badge is a milestone on that path. Excited to apply my advanced SQL skills to solve real-world challenges and contribute meaningfully to data-driven decision-making. Ready to bring my expertise to the table and drive success in any data-focused role.